[Score Plugin] Story 11336073 : - Ajout des xsd pour la 4.3.1                                 - Tests 3a sur power indicators updated                                 - Scores.xml updated

+++ src/libraries/3a/StaticModel.cpp
@@ -42,20 +42,18 @@ StaticModel::~StaticModel()
 // -----------------------------------------------------------------------------
 float StaticModel::ComputePower( const sword::UnitAttributes& message, const extractors::PowerExtractor_ABC& extractor ) const
 {
-    float res = 0;
-
+    double res = 0;
     for( int i = 0; i < message.equipment_dotations().elem().size(); ++i )
     {
         const sword::EquipmentDotations_EquipmentDotation equipmentMsg = message.equipment_dotations().elem( i );
         unsigned int powerValue = extractor.GetPowerValue( powerIndicators_->tools::Resolver< PowerIndicator, unsigned long >::Get( equipmentMsg.type().id() ) );
-        double equipmentRes = 0;        
+        double equipmentRes = 0;
         equipmentRes += equipmentMsg.available() * powerModificators_->GetAvailableModificator();
         equipmentRes += equipmentMsg.repairable() * powerModificators_->GetRepairableModificator();
         equipmentRes += equipmentMsg.repairing() * powerModificators_->GetRepairingModificator();
         equipmentRes += equipmentMsg.captured() * powerModificators_->GetCapturedModificator();
         equipmentRes += equipmentMsg.unavailable() * powerModificators_->GetUnavailableModificator();
-        res += static_cast< float >( equipmentRes * powerValue );
+        res += equipmentRes * powerValue;
     }
-
-    return res;
+    return static_cast< float >( res );
 }
